Prevent exception when payment has been cleared
Still haven't managed to spec this. Tried a controller test but couldn't get it to work. Consider system or unit test.

Note that in the real world, this avoids a crash, but still requires the user to click the button two more times before it will work, with no hints as to why. So not a great help.